TransientStepData

What is it?

TransientStepData objects hold information pertaining to transient sorption steps. This is dimensionless data that describes the transition from one equilibrium to another (e.g., concentration, dilation, etc) as a function of time.

How do I make one?

MembraneBase.TransientStepData — Type

TransientStepData(dataset)

Container for data concerning a single transient sorption step. I.e., a profile of mass concentration with time (starting from a previous equilibrium and generally reaching a new equilibrium) of some penetrant into an even film of polymer material, or more succinctly, a polymer slab.

Arguments

dataset: Should be some iterable with

the first item being an iterable of time data (in seconds)

the second item being an iterable of dimensionless sorption data (no units)

A number of functions exist to make your life easier with handling this data.

MembraneBase.resample — Method

resample(transient::TransientStepData, num_datapoints, time_function)

Resample the data with some time_function spacing and linear interpolation between adjacent data points. Use this method when you have a ridiculous amount of data and need to slim it down while losing as little valuable information as possible.

Arguments

transient::TransientStepData: TransientStepData used to resample

num_datapoints::Integer: number of resampled data points to be in the resulting TransientStepData (should always be less than what you currently have)

time_function::Symbol: how to weigh the time data...

:Linear will have even spacing from t=0 to t=num_datapoints

:Root will sample along the root of time (more information about the initial behavior will be kept)

It is usually best to use :Root for most scenarios and :Linear for quick scenarios

:Log will sample along the log_10 of time (much more info about initial behavior will be kept)

Reserve Log for extremely long timescales (think: on the order of months or years, not days or weeks)

This function assumes that time data is sorted in ascending order

MembraneBase.dataset — Method

Get a Vector of data in the transient step in form of [(time, sorption)...]. Used primarily for bootstrap-like resampling techniques.

MembraneBase.strip_measurement_to_value — Method

Return a copy of the data with no Measurement data types present.